Moved search tab to separate component
